complete all classes in order to have first results

+++ src/RdmGsaNet_pr08/generateNetEdge.java
@@ -9,8 +9,8 @@ public class generateNetEdge {
 		this.type = type ;
 	}
 	
-	public void generateEdge (int step) {
-		type.generateEdgeRule (step) ;
+	public void generateEdge ( double step ) {
+		type.generateEdgeRule ( step ) ;
 	}
 
 //--------------------------------------------------------------------------------------------------------------------
